requirements.txt
openai
numpy
matplotlib
Pillow
pyyaml
